upd main AST erro

+++ gcd.c
@@ -5,7 +5,7 @@ int gcd (int u, int v)
 }
    
 void main(void)
-{	int x; int y;
+{	float x; int y;
     x = input(); y = input();
     output(gcd(x,y));
 }
